+++ src/main/java/org/folio/notes/client/UsersClient.java
@@ -2,14 +2,14 @@ package org.folio.notes.client;
 
 import java.util.Optional;
 import org.folio.notes.domain.dto.User;
-import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.service.annotation.GetExchange;
+import org.springframework.web.service.annotation.HttpExchange;
 
-@FeignClient(value = "users", dismiss404 = true)
+@HttpExchange(url = "users", contentType = MediaType.APPLICATION_JSON_VALUE)
 public interface UsersClient {
 
-  @GetMapping(value = "/{id}", produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.TEXT_PLAIN_VALUE})
-  Optional<User> fetchUserById(@PathVariable("id") String id);
+  @GetExchange(value = "/{id}", accept = {MediaType.APPLICATION_JSON_VALUE, MediaType.TEXT_PLAIN_VALUE})
+  Optional<User> fetchUserById(@PathVariable String id);
 }