+++ src/main/java/org/folio/notes/domain/repository/NoteRepository.java
@@ -19,13 +19,13 @@ import org.folio.notes.domain.entity.NoteTypeEntity;
 import org.folio.notes.domain.entity.NoteTypeEntity_;
 import org.folio.spring.cql.JpaCqlRepository;
 import org.hibernate.query.sqm.internal.SqmCriteriaNodeBuilder;
+import org.jspecify.annotations.NonNull;
+import org.jspecify.annotations.Nullable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.lang.NonNull;
-import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Repository;
 
 @Repository