+++ NEWS.md
@@ -15,6 +15,7 @@
 
 ### Tech Dept
 * Refactor code to comply with Checkstyle method length limit (max 25 lines) ([MODNOTES-283](https://folio-org.atlassian.net/browse/MODNOTES-283))
+* Use GitHub workflow for CI/CD ([MODNOTES-291](https://folio-org.atlassian.net/browse/MODNOTES-291))
 
 ### Dependencies
 * Bump `LIB_NAME` from `OLD_VERSION` to `NEW_VERSION`